New unit test. Fixed issue in in HttpdCatalog.cc
