+++ modules/httpd_catalog_module/RemoteHttpResourceCache.cc
@@ -22,7 +22,7 @@
 //
 // You can contact OPeNDAP, Inc. at PO Box 112, Saunderstown, RI. 02874-0112.
 
-#include "config.h"
+#include <config.h>
 
 #include <sys/stat.h>
