+++ modules/httpd_catalog_module/unit-tests/test_config.h.in
@@ -4,5 +4,8 @@
 #define TEST_SRC_DIR "@abs_srcdir@"
 #define TEST_BUILD_DIR "@abs_builddir@"
 
+#define TEST_DATA_DIR "@abs_top_srcdir@/modules/httpd_catalog_module/data"
+const string test_data_dir = TEST_DATA_DIR;
+
 #endif
